requirements.txt
streamlit>=1.28.0
plotly>=5.17.0
pandas>=2.0.0
numpy>=1.24.0
firecrawl-py>=0.0.1
python-dotenv>=1.0.0
